Move `main` of `c2cpg` into `Main.scala`
