+++ joern-cli/frontends/c2cpg/src/main/scala/io/joern/c2cpg/passes/HeaderContentPass.scala
@@ -1,7 +1,7 @@
 package io.joern.c2cpg.passes
 
 import better.files.File
-import io.joern.c2cpg.C2Cpg.Config
+import io.joern.c2cpg.Main.Config
 import io.joern.c2cpg.datastructures.Global
 import io.joern.c2cpg.parser.FileDefaults
 import io.joern.c2cpg.utils.IncludeAutoDiscovery
