+++ joern-cli/frontends/c2cpg/src/test/scala/io/joern/c2cpg/fixtures/CpgAstOnlyFixture.scala
@@ -1,7 +1,7 @@
 package io.joern.c2cpg.fixtures
 
 import better.files.File
-import io.joern.c2cpg.C2Cpg.Config
+import io.joern.c2cpg.Main.Config
 import io.joern.c2cpg.passes.AstCreationPass
 import io.shiftleft.codepropertygraph.Cpg
