+++ joern-cli/frontends/c2cpg/src/test/scala/io/joern/c2cpg/passes/PreprocessorPassTests.scala
@@ -1,7 +1,7 @@
 package io.joern.c2cpg.passes
 
 import better.files.File
-import io.joern.c2cpg.C2Cpg.Config
+import io.joern.c2cpg.Main.Config
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
