+++ joern-cli/frontends/c2cpg/src/test/scala/io/joern/c2cpg/testfixtures/CCodeToCpgSuite.scala
@@ -1,7 +1,7 @@
 package io.joern.c2cpg.testfixtures
 
 import io.joern.c2cpg.C2Cpg
-import io.joern.c2cpg.C2Cpg.Config
+import io.joern.c2cpg.Main.Config
 import io.joern.c2cpg.parser.FileDefaults
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.semanticcpg.testfixtures.{CodeToCpgFixture, LanguageFrontend}
@@ -19,7 +19,7 @@ class C2CpgFrontend(override val fileSuffix: String = FileDefaults.C_EXT) extend
       outputPath = cpgFile.getAbsolutePath,
       includeComments = true
     )
-    c2cpg.runAndOutput(config)
+    c2cpg.createCpg(config)
   }
 }
